+++ assignment3/extension/united/Button.java
@@ -8,16 +8,16 @@ import java.awt.event.*;
 
 public class Button extends GameObject {
 
-	private int x;
-	private int y;
-	private int width;
-	private int height;
+	private double x;
+	private double y;
+	private double width;
+	private double height;
 	private String text = "";
 	
 	public GRect background;
 	public GLabel label;
 	
-	public Button(String text, int x, int y, int width, int height, GraphicsProgram canvas) {
+	public Button(String text, double x, double y, double width, double height, GraphicsProgram canvas) {
 		super(canvas);
 		this.text = text;
 		this.x = x;
@@ -34,16 +34,16 @@ public class Button extends GameObject {
 		object = collider;
 		
 		label = new GLabel(text);
-		final int XOffset = (width - (int)label.getWidth()) / 2;
-		final int YOffset = (height + (int)label.getAscent()) / 2;
+		final double XOffset = (width - label.getWidth()) / 2;
+		final double YOffset = (height + label.getAscent()) / 2;
 		label.setLocation(x+XOffset, y+YOffset);
 	}
 	
 	public void setLabel(String text) {
 		this.text = text;
 		label.setLabel(text);
-		final int XOffset = (width - (int)label.getWidth()) / 2;
-		final int YOffset = (height + (int)label.getAscent()) / 2;
+		final double XOffset = (width - label.getWidth()) / 2;
+		final double YOffset = (height + label.getAscent()) / 2;
 		label.setLocation(x+XOffset, y+YOffset);
 	}
