+++ assignment3/extension/united/Label.java
@@ -7,7 +7,7 @@ import java.awt.*;
 import java.awt.event.*;
 
 public class Label extends Button {
-	Label(String text, int x, int y, int width, int height, GraphicsProgram canvas) {
+	Label(String text, double x, double y, double width, double height, GraphicsProgram canvas) {
 		super(text, x, y, width, height, canvas);
 		background.setVisible(false);
 		object.setVisible(false);
